Removed colors value casting to short Closes #1354

+++ rt4/CacheObjectConfig.java
@@ -29,7 +29,7 @@ class CacheObjectConfig {
 	public int stageOperationId = -1;
 	public int stageIndex = -1;
 	public int zTranslate = 0;
-	public short[] originalColors, modifiedColors;
+	public int[] originalColors, modifiedColors;
 
 	public CacheObjectConfig(final Block.Sector sector, final int index) {
 		this.index = index;
@@ -129,11 +129,11 @@ class CacheObjectConfig {
 				break;
 			case 40: {
 				final int len = stream.getUByte();
-				originalColors = new short[len];
-				modifiedColors = new short[len];
+				originalColors = new int[len];
+				modifiedColors = new int[len];
 				for (int i = 0; i < len; i++) {
-					originalColors[i] = (short) stream.getUShort();
-					modifiedColors[i] = (short) stream.getUShort();
+					originalColors[i] = stream.getUShort();
+					modifiedColors[i] = stream.getUShort();
 				}
 				break;
 			}
